Update doDelete()
in case of key is empty, doDelete trys to delete "/" and fails because it's directory

+++ src/Sesshin/Storage/Files.php
@@ -82,7 +82,7 @@ class Files extends Storage {
 	
 	protected function doDelete($key) {
     $file = $this->getPath().'/'.$key;
-    if (file_exists($file)) {
+    if (file_exists($file) && is_file($file)) {
       return unlink($file);
     }
     return false;
